restructure the legend files to make it cleaner and discoverable and update the imports
