+++ src/charts/MultiLineChart.js
@@ -6,7 +6,7 @@ import { scaleBand, scaleLinear } from "d3-scale";
 import { axisLeft } from "d3-axis";
 import BaseChart from "./BaseChart";
 import { loadColumnData } from "@/datastore/decorateColumnMethod";
-import { buildColorLegendSpec } from "@/react/colorLegend/buildColorLegendSpec";
+import { buildColorLegendSpec } from "@/react/legend/color_legend/buildColorLegendSpec";
 
 class MultiLineChart extends SVGChart {
     constructor(dataStore, div, config) {
