+++ src/charts/SingleHeatMap.js
@@ -2,7 +2,7 @@ import { select } from "d3-selection";
 import { easeLinear } from "d3-ease";
 import BaseChart from "./BaseChart";
 import SVGChart from "./SVGChart.js";
-import { buildColorLegendSpec } from "@/react/colorLegend/buildColorLegendSpec";
+import { buildColorLegendSpec } from "@/react/legend/color_legend/buildColorLegendSpec";
 
 class SingleHeatMap extends SVGChart {
     constructor(dataStore, div, config) {
