+++ src/react/legend/color_legend/buildColorLegendSpec.ts
