+++ src/react/legend/shared/legendConstants.ts
