+++ src/react/legend/shared/legendTypes.ts
