+++ src/tests/react/legend/buildColorLegendSpec.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, test, vi } from "vitest";
 import DataStore from "@/datastore/DataStore";
-import { buildColorLegendSpec } from "./buildColorLegendSpec";
+import { buildColorLegendSpec } from "@/react/legend/color_legend/buildColorLegendSpec";
 
 function createStore(
     columns: Record<string, Record<string, unknown>>,
